+++ fltk-sys/build.rs
@@ -97,6 +97,10 @@ fn main() {
         dst.define("OpenGL_GL_PREFERENCE", "GLVND");
     }
 
+    if cfg!(feature = "gl-window") {
+        dst.define("FLTK_GL_WIN_SUPPORT", "ON");
+    }
+
     if cfg!(feature = "cpp-testing") {
         println!("cargo:rerun-if-changed=cfltk/tests/test1.cpp");
         dst.define("CFLTK_BUILD_TESTS", "ON");
@@ -136,7 +140,9 @@ fn main() {
     if !cfg!(feature = "fltk-shared") {
         println!("cargo:rustc-link-lib=static=fltk");
         println!("cargo:rustc-link-lib=static=fltk_images");
-        // println!("cargo:rustc-link-lib=static=fltk_gl");
+        if cfg!(feature = "gl-window") {
+            println!("cargo:rustc-link-lib=static=fltk_gl");
+        }
 
         if cfg!(feature = "system-libpng") {
             println!("cargo:rustc-link-lib=dylib=png");
@@ -162,7 +168,9 @@ fn main() {
                 println!("cargo:rustc-link-lib=framework=Carbon");
                 println!("cargo:rustc-link-lib=framework=Cocoa");
                 println!("cargo:rustc-link-lib=framework=ApplicationServices");
-                println!("cargo:rustc-link-lib=framework=OpenGL");
+                if cfg!(feature = "gl-window") {
+                    println!("cargo:rustc-link-lib=framework=OpenGL");
+                }
             }
             "windows" => {
                 if cfg!(target_env = "gnu") {
@@ -181,8 +189,10 @@ fn main() {
                 println!("cargo:rustc-link-lib=dylib=user32");
                 println!("cargo:rustc-link-lib=dylib=kernel32");
                 println!("cargo:rustc-link-lib=dylib=odbc32");
-                println!("cargo:rustc-link-lib=dylib=opengl32");
-                println!("cargo:rustc-link-lib=dylib=glu32");
+                if cfg!(feature = "gl-window") {
+                    println!("cargo:rustc-link-lib=dylib=opengl32");
+                    println!("cargo:rustc-link-lib=dylib=glu32");
+                }
             }
             _ => {
                 println!("cargo:rustc-link-lib=dylib=stdc++");
@@ -195,7 +205,9 @@ fn main() {
                 println!("cargo:rustc-link-lib=dylib=Xfixes");
                 println!("cargo:rustc-link-lib=dylib=Xft");
                 println!("cargo:rustc-link-lib=dylib=fontconfig");
-                println!("cargo:rustc-link-lib=dylib=glx");
+                if cfg!(feature = "gl-window") {
+                    println!("cargo:rustc-link-lib=dylib=GL");
+                }
             }
         }
     }