+++ src/lib.rs
@@ -177,6 +177,7 @@
 //! - system-libjpeg: Uses the system libjpeg
 //! - system-zlib: Uses the system zlib
 //! - legacy-opengl: Support of Lagacy OpenGL, the crate uses GLVND by default
+//! - gl-window: Support for GlWindow
 //!
 //! ## FAQ
 //!