+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 ### Changes
-- Add GlWindow.
+- Add GlWindow under gl-window feature flag.
 
 ## [0.4.6] - 2020-05-20
 ### Changes